requirements.txt
py-clob-client>=0.24.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
requests>=2.31.0
